Change QR code size in withdraw , change change number type that has E to long , issue #328

+++ app/src/main/java/de/netalic/falcon/ui/addresses/AddressesRecyclerViewAdapter.java
@@ -49,7 +49,7 @@ public class AddressesRecyclerViewAdapter extends RecyclerView.Adapter<RecyclerV
         WalletViewHolder walletViewHolder = (WalletViewHolder) holder;
         Wallet wallet = mWalletList.get(position);
         walletViewHolder.mTextViewWalletName.setText(wallet.getName());
-        walletViewHolder.mTextViewBalance.setText(wallet.getCurrencySymbol() + " " + String.valueOf(wallet.getBalance()));
+        walletViewHolder.mTextViewBalance.setText(wallet.getCurrencySymbol() + " " + String.valueOf(Double.valueOf(wallet.getBalance()).longValue()));
         walletViewHolder.mTextViewAddress.setText(wallet.getAddress());
 
     }
