+++ app/src/main/java/de/netalic/falcon/ui/transfer/TransferFragment.java
@@ -109,7 +109,7 @@ public class TransferFragment extends Fragment implements TransferContract.View 
             @Override
             public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
 
-                mTextViewBalance.setText(String.valueOf(mWalletList.get(position).getBalance()) + " " + mWalletList.get(position).getCurrencySymbol());
+                mTextViewBalance.setText(String.valueOf(Double.valueOf(mWalletList.get(position).getBalance()).longValue()) + " " + mWalletList.get(position).getCurrencySymbol());
                 mPosition = position;
             }
